backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
python-dotenv
httpx
alembic
